backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
httpx==0.25.2
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiosmtplib==3.0.1
firebase-admin==6.2.0
pymysql==1.1.0
cryptography==41.0.7
